Add files via upload
In this project, we'll work with a dataset of used cars from eBay Kleinanzeigen, a classifieds section of the German eBay website.  
The dataset was originally scraped and uploaded to [Kaggle]('https://www.kaggle.com/orgesleka/used-cars-database/data'). For the sake of the project few modifications have been done from the original dataset that was uploaded to Kaggle.
**The aim of this project is to clean the data and analyze the included used car listings.**